Move platform interfaces to TabletDriverPlugin
