--- TabletDriverLib/Interop/Converters/IConverter.cs
@@ -1,8 +1,0 @@
-﻿namespace TabletDriverLib.Interop.Converters
-{
-    public interface IConverter<T1, T2>
-    {
-        T1 Convert(T2 obj);
-        T2 Convert(T1 obj);
-    }
-}
