+++ TabletDriverLib/Interop/Display/XScreen.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using System.Runtime.InteropServices;
 using NativeLib.Linux.Xorg;
 using TabletDriverPlugin;
+using TabletDriverPlugin.Platform.Display;
 
 namespace TabletDriverLib.Interop.Display
 {
