+++ TabletDriverLib/Output/AbsoluteMode.cs
@@ -1,9 +1,9 @@
 using System.Collections.Generic;
 using System.Linq;
 using TabletDriverLib.Interop;
-using TabletDriverLib.Interop.Cursor;
 using TabletDriverPlugin;
 using TabletDriverPlugin.Attributes;
+using TabletDriverPlugin.Platform.Pointer;
 using TabletDriverPlugin.Tablet;
 
 namespace TabletDriverLib.Output
