+++ TabletDriverLib/Output/RelativeMode.cs
@@ -2,9 +2,9 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using TabletDriverLib.Interop;
-using TabletDriverLib.Interop.Cursor;
 using TabletDriverPlugin;
 using TabletDriverPlugin.Attributes;
+using TabletDriverPlugin.Platform.Pointer;
 using TabletDriverPlugin.Tablet;
 
 namespace TabletDriverLib.Output
