+++ TabletDriverPlugin/Platform/Display/IDisplay.cs
@@ -1,6 +1,6 @@
 using TabletDriverPlugin;
 
-namespace TabletDriverLib.Interop.Display
+namespace TabletDriverPlugin.Platform.Display
 {
     public interface IDisplay
     {
