+++ TabletDriverPlugin/Platform/Pointer/MouseButton.cs
@@ -1,4 +1,4 @@
-namespace TabletDriverLib.Interop.Cursor
+namespace TabletDriverPlugin.Platform.Pointer
 {
     public enum MouseButton
     {
